+++ CHANGELOG.md
@@ -4,10 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - ???
+### New
+- Atari Wireless Joystick: Add support for it
+
+### Changed
+- Allowlist: rename `uni_bt_allowlist_allow_addr()` to `uni_bt_allowlist_is_allowed_addr()`
+
 ## [3.10.2] - 2023-11-13
 ### Changed
 - BTstack: Upgraded to latest develop branch as of 2023-11-13
-- Allowlist: rename `uni_bt_allowlist_allow_addr()` to `uni_bt_allowlist_is_allowed_addr()`
 
 ### Fixed
 - Xbox: Axis works Ok (fixed in latest BTstack)